Split App.js into model view; refactored CSS
