+++ src/About/About.css
@@ -5,10 +5,13 @@
 
 .about-content {
   position: relative;
-  padding-bottom: 0.5rem;
-
   display: grid;
   grid-template-rows: auto 1fr;
+
+  box-shadow: 8px 8px 8px rgba(0, 0, 0, 0.15);
+  background: var(--neutralLight);
+  padding: 1.5rem;
+  padding-bottom: 0.5rem;
 }
 
 .about-content::before,
@@ -57,7 +60,7 @@
   height: 34px;
   padding: 0 5px;
   opacity: 0.33;
-  transition: all var(--animationDuration) var(--animationEasing);
+  transition: all var(--transitionDuration) var(--transitionEasing);
   transition-property: opacity, transform;
 }
