+++ src/App/App.css
@@ -1,0 +1,8 @@
+.fade-in {
+  transition: opacity var(--transitionDuration) var(--transitionEasing);
+  opacity: 0;
+}
+
+.fade-in-start {
+  opacity: 1;
+}
