+++ src/Article/Article.css
@@ -5,7 +5,7 @@
 .article-item-title {
   font-size: 22px;
   color: var(--foregroundDark);
-  transition: color var(--animationDuration) var(--animationEasing);
+  transition: color var(--transitionDuration) var(--transitionEasing);
 }
 
 .article-item-title:hover {
