+++ src/Page/Page.css
@@ -14,6 +14,6 @@
   left: 0;
   /* Make background white for a beautiful page transition */
   background: var(--neutralLighter);
-  transition: opacity var(--animationDuration) var(--animationEasing);
+  transition: opacity var(--transitionDuration) var(--transitionEasing);
   opacity: 0;
 }
