--- src/animation.css
@@ -1,13 +1,0 @@
-.fade-in {
-  transition: opacity var(--animationDuration) var(--animationEasing);
-  opacity: 0;
-}
-
-.fade-in-start {
-  opacity: 1;
-}
-
-@keyframes spinRight {
-  from { transform: rotate(0deg); }
-  to { transform: rotate(360deg); }
-}
