+++ src/index.js
@@ -1,6 +1,3 @@
 import './favicon.png';
-
-import './animation.css';
 import './style.css';
-
 import './App';
